fix: Let OutputBuilders do their stuff with the ResponseBag and not write messy headers
The method `getHeaders` of output builders return either the array of headers of the ResponseBbag or an array of string with
headers, the response headers was messy:
```
export.csv:
no:
chunked:
application/octet-stream:
X-Sendfile: export.csv
```

It's okay with HTTP/1.1 but not HTTP/2.

Headers are already written in \Pyrite\Stack\Application::buildResponseFromResponseBag using the ResponesBag headers.

Breaking change: OutputBuilder interface changed.